sittir-core/napi: \$nodeId is tree-sitter Node::id() — drop synthetic counter
Was: \`sittir_core::read_node\` assigned a synthetic monotonic \`u32\`
counter as \`\$nodeId\` and threaded \`&mut next_node_id\` through every
call. JS-side \`TreeHandle.nodeById\` consumed tree-sitter's pointer-
derived \`Node::id()\` instead. Two id spaces, no cross-engine drill-
in.

Now: \`\$nodeId\` is \`Some(node.id() as u64)\` — tree-sitter's
canonical per-tree node identifier. Same value JS receives from the
wasm parser. Counter + threading parameter dropped.

Files:
  - rust/crates/sittir-core/src/types.rs:
      NodeData.node_id: Option<u32> → Option<u64>
  - rust/crates/sittir-core/src/read_node.rs:
      read_node(tree, source, node_id) — drops next_node_id param
      find_by_id walks by node.id() match instead of pre-order index
  - rust/crates/sittir-{rust,typescript,python}-napi/src/lib.rs:
      SittirEngine drops next_node_id field
      parse_and_read / read_node call read_node(... 3 args)
      read_node(node_id: f64) — JS number is the natural carrier
        for u64 (53-bit precision covers user-space pointers)
      panic_msg helper extracted (was inline boilerplate ×2)
  - codegen/src/emitters/wrap.ts: dispatch logic in readNode honors
    the new id source

Drill-in caveat documented in the emitter: tree-sitter Node::id()
is PER-TREE, and the wasm parser + napi engine parse two DISTINCT
trees with two DISTINCT id spaces. Passing a wasm-side nodeId to
engine.readNode(...) panics with "node id N not found in current
tree". Wrap.ts's drill-in still falls back to JS until the parse is
unified — proper fix is having the JS-side TreeHandle wrap the napi
engine's tree (one parse, one reader). Inline comment explains.

Whole workspace cargo-checks; all 3 napi crates rebuilt; counts hold:
  rust:       from 130/148, rt 122/136, factory 424/965
  typescript: from 127/137, rt  96/112, factory 387/916
  python:     from 107/114, rt 105/115, factory 205/860

+++ rust/crates/sittir-core/src/types.rs
@@ -68,12 +68,20 @@ pub struct NodeData {
     )]
     pub span: Option<Span>,
 
+    /// Tree-sitter's canonical node id (from `Node::id()`, a pointer-
+    /// derived `usize`). Identical id-space on both engines: JS-side
+    /// `TreeHandle.nodeById` and the napi `readNode($nodeId)` consume
+    /// the same value. Was a synthetic monotonic counter pre-fix —
+    /// that invented a parallel id-space and broke drill-in dispatch
+    /// across engines. Serialized as a JSON number; macOS / Linux
+    /// user-space pointers fit in 53 bits so the f64 round-trip is
+    /// exact.
     #[serde(
         rename = "$nodeId",
         default,
         skip_serializing_if = "Option::is_none"
     )]
-    pub node_id: Option<u32>,
+    pub node_id: Option<u64>,
 }
 
 /// Where a `NodeData` originated. `Ts` = `readNode` over a tree-sitter
